Extract operational API router
