+++ src/app/routers/__init__.py
@@ -1,0 +1,1 @@
+"""API router modules for lotus-risk."""
